fix(seo): keep docs page exports valid
