+++ web/src/app/docs/[[...slug]]/page.tsx
@@ -11,24 +11,11 @@ import {
   getDocBySlug,
   getDocNeighbors,
 } from "@/lib/docs";
+import { docsSchemaId } from "../docs-schema-id";
 
 type Props = {
   params: Promise<{ slug?: string[] }>;
 };
-
-export function docsSchemaId(href: string) {
-  const pathname = href
-    .replace(/^https?:\/\/[^/]+/, "")
-    .replace(/\/+$/, "");
-  const slug = pathname
-    .replace(/^\/docs(?:\/|$)/, "")
-    .replace(/^\/+|\/+$/g, "")
-    .replace(/\//g, "-");
-
-  return `agentclash-docs-${
-    slug || "home"
-  }-schema`;
-}
 
 export function generateStaticParams() {
   return getAllDocSlugs().map((slug) => ({ slug }));
